+++ repentogon/LuaInterfaces/Room/LuaRoomConfig.cpp
@@ -1,7 +1,7 @@
 #include "IsaacRepentance.h"
 #include "LuaCore.h"
 #include "HookSystem.h"
-#include "../../Patches/VirtualRooms.h"
+#include "../../Patches/VirtualRoomSets.h"
 
 /*LUA_FUNCTION(Lua_GameGetRoomConfig) {
 	Game* game = lua::GetRawUserdata<Game*>(L, 1, lua::Metatables::GAME, "Game");
