+++ repentogon/LuaInterfaces/Room/LuaRoomConfigSet.cpp
@@ -1,7 +1,7 @@
 #include "IsaacRepentance.h"
 #include "LuaCore.h"
 #include "HookSystem.h"
-#include "../../Patches/VirtualRooms.h"
+#include "../../Patches/VirtualRoomSets.h"
 
 LUA_FUNCTION(Lua_RoomConfigSetGetRoom)
 {
